+++ etc.h
@@ -1,0 +1,20 @@
+//
+// Created by eran on 18/08/18.
+//
+
+#ifndef FINAL_ETC_H
+#define FINAL_ETC_H
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <printf.h>
+#include "gameUtils.h"
+
+
+gameParams *getExampleGame(int n, int m) ;
+
+
+
+
+
+#endif //FINAL_ETC_H
